zebra: No need to send `ack` of initial registration for nht
When we receive a request from an upper level protocol we
are sending a response to the protocol immediately( before
lookup and handling ) that has no nexthops.

This is interesting/bad for two reasons:

1) There is no code path where we don't send two responses for
the initial registration
2) BGP was taking this information and resetting the connection attempt
causing the first connection attempt to pretty much always go into
connection time out.  The connection timeout for bgp is 2 minutes
in most cases!

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>