bgpd: When we receive a try this command again respect it
We have setup non-blocking connect.  It is possible for the
connect to not be finished and for the socket system to tell
us to try this command again.  The getsockopt call was
not respecting this request and we were causing the Connect
to fail.  Since the connect failed, this would cause the
connect timer to be started, which by default is 2 minutes.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>