doc: add "Getting Started" section
While we have docs on various pieces of the build system we don't have
any docs on how to actually get FRR running once it's installed, nor do
we have comprehensive documentation on the basic procedure for building
from source. This patch remedies both of those.

Also updated the services list in the docs and removed the SERVICES file
from the project root.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>