bgpd: Do not set interfaces ifindex
When a interface is deleted we should not be setting
the interface's ifindex to IFINDEX_INTERNAL.  This is
causing issues when we get new messages from zebra
about what to do with this interface.

I believe this is a hold-over from the very first vrf
implementation and we were not properly handling
interface movements and were allowing the daemons to
keep some of the state when a call was received.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>