bgpd: Cleanup some duplicated code
We have 2 code paths that were duplicating a bunch of code
for the deletion of connected prefixes.

This simplifies the code path and makes the code look a bit
cleaner.

I did not touch the _add path because the v4 if statement
had some code I did not have time to look into.  Future project.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>