ldpd: Schedule application of config till after read-in
With commit e94b38d94b5 we are now scheduling the read
of vty config until after the startup of main thread
processing.  It now becomes necessary to move the
application of the config until after the read
in of the config from a file if we are using a
non-integrated config.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>